refactor: change the bridge name to NativeRudderBridge in iOS
This is to avoid any naming conflict issue in iOS
