+++ libs/sdk/ios/RNRudderSdkModule.mm
@@ -14,7 +14,7 @@ RCT_EXPORT_MODULE()
 - (std::shared_ptr<facebook::react::TurboModule>)getTurboModule:
     (const facebook::react::ObjCTurboModule::InitParams &)params
 {
-    return std::make_shared<facebook::react::NativeBridgeSpecJSI>(params);
+    return std::make_shared<facebook::react::NativeRudderBridgeSpecJSI>(params);
 }
 
 + (BOOL)requiresMainQueueSetup
